Remove legacy colors and fix import paths
